requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
django==4.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.3.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
psycopg==3.1.12
    # via -r requirements.in
psycopg-binary==3.1.12
    # via -r requirements.in
psycopg-pool==3.1.8
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.8.0
    # via
    #   asgiref
    #   psycopg
    #   psycopg-pool
